#  CachedTimeSource - Added Thread.MemoryBarrier to avoid code-reordering