- [x] Verify performance with Benchmark.NET (NET Framework 4.8 + NET6 + force 32bit + force 64bit)