@snakefoot 2023-12-02
Before:
```
| Method                  | Job        | Platform | Mean       | Error     | StdDev    | Rank | Allocated |
|------------------------ |----------- |--------- |-----------:|----------:|----------:|-----:|----------:|
| FastUtcTimeSource       | Job-JWMLCL | X64      |   4.159 ns | 0.0277 ns | 0.0260 ns |    1 |         - |
| FastLocalTimeSource     | Job-JWMLCL | X64      |   4.688 ns | 0.0306 ns | 0.0286 ns |    2 |         - |
| FastLocalTimeSource     | Job-BNYUHE | X86      |   5.910 ns | 0.0590 ns | 0.0523 ns |    3 |         - |
| FastUtcTimeSource       | Job-BNYUHE | X86      |   5.912 ns | 0.0326 ns | 0.0289 ns |    3 |         - |
| AccurateUtcTimeSource   | Job-JWMLCL | X64      |  67.685 ns | 0.4201 ns | 0.3508 ns |    4 |         - |
| AccurateUtcTimeSource   | Job-BNYUHE | X86      | 134.895 ns | 1.0986 ns | 1.0277 ns |    5 |         - |
| AccurateLocalTimeSource | Job-JWMLCL | X64      | 276.942 ns | 1.6234 ns | 1.2674 ns |    6 |         - |
| AccurateLocalTimeSource | Job-BNYUHE | X86      | 780.794 ns | 4.7053 ns | 4.4013 ns |    7 |         - |
```
After:
```
| Method                  | Job        | Platform | Mean       | Error     | StdDev    | Rank | Allocated |
|------------------------ |----------- |--------- |-----------:|----------:|----------:|-----:|----------:|
| FastLocalTimeSource     | Job-JWMLCL | X64      |   5.967 ns | 0.0547 ns | 0.0512 ns |    1 |         - |
| FastUtcTimeSource       | Job-JWMLCL | X64      |   7.206 ns | 0.0421 ns | 0.0393 ns |    2 |         - |
| FastLocalTimeSource     | Job-BNYUHE | X86      |  21.429 ns | 0.0661 ns | 0.0618 ns |    3 |         - |
| FastUtcTimeSource       | Job-BNYUHE | X86      |  21.454 ns | 0.0917 ns | 0.0858 ns |    3 |         - |
| AccurateUtcTimeSource   | Job-JWMLCL | X64      |  65.989 ns | 0.3348 ns | 0.2968 ns |    4 |         - |
| AccurateUtcTimeSource   | Job-BNYUHE | X86      | 130.774 ns | 0.4283 ns | 0.3344 ns |    5 |         - |
| AccurateLocalTimeSource | Job-JWMLCL | X64      | 267.692 ns | 1.1005 ns | 1.0294 ns |    6 |         - |
| AccurateLocalTimeSource | Job-BNYUHE | X86      | 759.154 ns | 3.7541 ns | 3.5116 ns |    7 |         - |
```
So 4 times slower performance-hit when 32bit / x86