@snakefoot 2023-12-03
Seems that Intel CPU uses the FPU-registers for handling 64-bit-values, thus even when running in 32-bit-mode then the chance of tearing is minimal.