@snakefoot 2023-12-10
Instead performed new benchmark of adding `MemoryBarrier` to ensure `_lastTime` is always assigned before `_lastTicks`.

| Method                  | Job          | Jit       | Platform | Mean       | Error      | StdDev     |
|------------------------ |------------- |---------- |--------- |-----------:|-----------:|-----------:|
| FastUtcTimeSource       | RyuJitX64    | RyuJit    | X64      |   3.611 ns |  0.0622 ns |  0.0582 ns |      
| FastUtcTimeSource       | LegacyJitX86 | LegacyJit | X86      |   5.818 ns |  0.0958 ns |  0.0896 ns |    
| FastLocalTimeSource     | RyuJitX64    | RyuJit    | X64      |   3.590 ns |  0.0673 ns |  0.0629 ns |         
| FastLocalTimeSource     | LegacyJitX86 | LegacyJit | X86      |   6.047 ns |  0.1187 ns |  0.1052 ns |
| PureDateTimeUtcNow          | RyuJitX64    | RyuJit    | X64      |  67.222 ns |  1.0602 ns |  0.9918 ns |       
| PureDateTimeUtcNow          | LegacyJitX86 | LegacyJit | X86      | 136.818 ns |  2.1062 ns |  1.9702 ns | 
| PureDateTimeLocalNow        | RyuJitX64    | RyuJit    | X64      | 271.983 ns |  4.2370 ns |  3.9633 ns |     
| PureDateTimeLocalNow        | LegacyJitX86 | LegacyJit | X86      | 777.621 ns | 13.4024 ns | 12.5367 ns |    
| AccurateUtcTimeSource   | RyuJitX64    | RyuJit    | X64      |  71.206 ns |  1.4384 ns |  2.9706 ns |   
| AccurateUtcTimeSource   | LegacyJitX86 | LegacyJit | X86      | 138.488 ns |  2.6541 ns |  2.4827 ns |   
| AccurateLocalTimeSource | RyuJitX64    | RyuJit    | X64      | 272.284 ns |  5.4456 ns |  5.0938 ns | 
| AccurateLocalTimeSource | LegacyJitX86 | LegacyJit | X86      | 779.333 ns | 14.0040 ns | 13.0993 ns |